% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LNConditionalMeans.R
\name{left_censored_loglik}
\alias{left_censored_loglik}
\title{Calculate the log likelihood of right censored lognormal data.}
\usage{
left_censored_loglik(params, cc, flg)
}
\arguments{
\item{params}{A two item list or vector containing the mean and sd of the
distribution on the log scale (i.e., the parameters of the related
normal distribution).}

\item{cc}{A vector of data values, including both observed values, where they
exist, or the detection limits, where data was censored.}

\item{flg}{A vector of TRUE or FALSE values, of the same length as cc, that
indicates which values are detection limits (TRUE) and which are measured
values (FALSE)}
}
\value{
The calculated log likelihood under the assumption of a lognormal
 distribution.
}
\description{
Calculate the log likelihood of right censored lognormal data.  The
log likelihood has two parts -- a term for observed values over the
detection limit, and term for censored values (where all we know is that
the value falls below the detection limit).
}
\examples{
df = data.frame(sim = sort(rlnorm(25,2,3)),
                  cens=c(rep.int(TRUE,5), rep.int(FALSE,20)))
df$sim[1:4] <- df$sim[5]
left_censored_loglik(c(2,5), df$sim, df$cens)
left_censored_loglik(c(2,3), df$sim, df$cens)
left_censored_loglik(c(1,3), df$sim, df$cens)
}
