% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LNConditionalMeans.R
\name{simulate_mean_censored}
\alias{simulate_mean_censored}
\title{Estimate the mean of unobserved left censored values}
\usage{
simulate_mean_censored(lmu, lsigma, cutoff, sz = 1000)
}
\arguments{
\item{lmu}{The mean (on the log scale) of the lognormal distribution.}

\item{lsigma}{The standard deviation of the lognormal distribution.}

\item{cutoff}{Vector of detection limits (or observed values).}

\item{sz}{The target size of the (simulated) sample for calculating
conditional means (default = 1000). The current approach simulates a large
draw from the underlying uncensored lognormal distribution, and retains
only values below the detection limits.  The function estimates the size of
the oversample needed, but because this is based on probabalistic reasoning,
the sample is not guaranteed to always be exactly the size requested.}
}
\value{
a vector of estimated conditional means.  Note that this function
provides conditional means for all observations, not only the censored ones.
This is wasteful, and potentially confusing to users.  See examples.
}
\description{
\code{simulate_mean_censored} estimates conditional means of
values from specified lognormal distribution (with specific parameters) that
fall below (variable) cutoff values.  Estimated means are determined via
simulation.
}
\details{
This functions in inefficient in the current context, as it simulates means
for all values, not just censored values.  That will make it slow for large
data sets or high values of sz.
}
\examples{
df <- data.frame(sim = sort(stats::rlnorm(25,2,3)),
                cens=c(rep.int(TRUE,5), rep.int(FALSE,20)))
df$sim[1:4] <- df$sim[5]
est <- simulate_mean_censored(lmu = 2, lsigma = 3, cutoff = df$sim)
library(ggplot2)
ggplot(df, aes(x = 1:25)) +
geom_line(aes(y = sim, color=cens)) +
  geom_point(aes(y = est)) +
  scale_y_log10() +
  scale_color_discrete(name = 'Censored') +
  theme_minimal() +
  xlab('Rank Order') +
  ylab('Raw Data (Line) and Conditional Means (points)')
rm(est)
}
